moved logger defn to files.py
